Unifies Actor, ActorTemplate and ActorTemplateVersion resource refs
